Rename image series annotator to batch annotator
