+++ micro_sam/sam_annotator/__init__.py
@@ -3,5 +3,5 @@
 
 from .annotator import annotator, Annotator
 from .annotator_tracking import annotator_tracking
-from .image_series_annotator import image_folder_annotator, image_series_annotator
+from .batch_annotator import image_folder_annotator, image_series_annotator
 # from .object_classifier import object_classifier
