+++ setup.cfg
@@ -78,7 +78,7 @@ napari.manifest =
 console_scripts =
     micro_sam.annotator = micro_sam.sam_annotator.annotator:main
     micro_sam.annotator_tracking = micro_sam.sam_annotator.annotator_tracking:main
-    micro_sam.image_series_annotator = micro_sam.sam_annotator.image_series_annotator:main
+    micro_sam.image_series_annotator = micro_sam.sam_annotator.batch_annotator:main
     micro_sam.precompute_embeddings = micro_sam.precompute_state:main
     micro_sam.automatic_segmentation = micro_sam.v1.automatic_segmentation:main
     micro_sam.train = micro_sam.v1.training.training:main
